+++ Krowi_AchievementFilter/Data/ExportedNextPatchAchievements.lua
@@ -1,4 +1,4 @@
--- [[ Exported at 2021-11-14 00-59-13 ]] --
+-- [[ Exported at 2021-11-14 21-03-12 ]] --
 -- [[ This code is automatically generated as an export from ]] --
 -- [[ an SQLite database and is not meant for manual edit. ]] --